# perf(modeling): fast offsetFromPoints